chore: extract shared from dashboard index (move-only)
